+++ tests/got/test_net_connect.js
@@ -10,7 +10,7 @@ const servers = require('../servers')
 
 describe('`disableNetConnect()`', () => {
   // TODO: https://github.com/mswjs/interceptors/issues/474
-  it.skip('prevents connection to unmocked hosts', async () => {
+  it('prevents connection to unmocked hosts', async () => {
     nock.disableNetConnect()
 
     nock('http://www.example.test').get('/').reply(200)
@@ -22,7 +22,7 @@ describe('`disableNetConnect()`', () => {
   })
 
   // TODO: https://github.com/mswjs/interceptors/issues/474
-  it.skip('prevents connections when no hosts are mocked', async () => {
+  it('prevents connections when no hosts are mocked', async () => {
     nock.disableNetConnect()
 
     await assertRejects(got('http://example.test'), err => {
@@ -52,7 +52,7 @@ describe('`enableNetConnect()`', () => {
   })
 
   // TODO: https://github.com/mswjs/interceptors/issues/474
-  it.skip('disallows request for other domains, via string', async () => {
+  it('disallows request for other domains, via string', async () => {
     nock.enableNetConnect('localhost')
 
     await assertRejects(
@@ -76,7 +76,7 @@ describe('`enableNetConnect()`', () => {
   })
 
   // TODO: https://github.com/mswjs/interceptors/issues/474
-  it.skip('disallows request for other domains, via regexp', async () => {
+  it('disallows request for other domains, via regexp', async () => {
     nock.enableNetConnect(/ocalhos/)
 
     await assertRejects(
@@ -100,7 +100,7 @@ describe('`enableNetConnect()`', () => {
   })
 
   // TODO: https://github.com/mswjs/interceptors/issues/474
-  it.skip('disallows request for other domains, via function', async () => {
+  it('disallows request for other domains, via function', async () => {
     nock.enableNetConnect(host => host.includes('ocalhos'))
 
     await assertRejects(
@@ -110,7 +110,7 @@ describe('`enableNetConnect()`', () => {
   })
 
   // TODO: https://github.com/mswjs/interceptors/issues/474
-  it.skip('passes the domain to be tested, via function', async () => {
+  it('passes the domain to be tested, via function', async () => {
     const matcher = sinon.stub().returns(false)
     nock.enableNetConnect(matcher)